+++ CHANGELOG.md
@@ -6,12 +6,12 @@
 ## [1.1.0] - 2017-07-28
 ### Added
  - Support for `${definition...}` placeholders in glob patterns for completion item providers.
+ - Support hot configuration reload.
 
 ## [1.0.0] - 2017-07-26
 ### Added
  - Completion item providers.
  - Definition providers.
-
 
 [Unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/v1.1.0...HEAD
 [1.1.0]: https://github.com/olivierlacan/keep-a-changelog/compare/v1.1.0...v1.0.0